Fix frozen norm weights: keep model fp32 for optimizer precision
Model was cast to bf16 before accelerator.prepare(), causing the
optimizer to work on bf16 parameters directly. In bfloat16, the
smallest representable change to 1.0 is ±0.0078 — but AdamW updates
at lr=2e-4 are ~0.0002, rounding to zero. All 113 normalization
layers (RMSNorm, QK LayerNorm) were frozen at initialization.

Fix: let Accelerate handle mixed precision (forward/backward in bf16,
optimizer states in fp32). Verified locally that all norm weights
now update correctly.

+++ scripts/train.py
@@ -307,10 +307,10 @@ def main():
         model = Qwen3ForCausalLM.from_pretrained(
             resume_dir,
             attn_implementation="flash_attention_2",
-            torch_dtype=torch.bfloat16,
+            torch_dtype=torch.float32,
         )
     else:
-        model = Qwen3ForCausalLM(config).to(torch.bfloat16)
+        model = Qwen3ForCausalLM(config)
 
     model.gradient_checkpointing_enable()
